rework structor for testing
