+++ utils/types/responses.ts
@@ -1,13 +1,17 @@
 import { Attribute, Category, InventoryItem, ItemDefinition, User } from '.'
 
-export type UserResponse = User
+export interface UserResponse extends User {
+  _id: string
+}
 
 export interface ItemDefinitionResponse extends ItemDefinition {
+  _id: string
   category?: Category
   attributes?: Attribute[]
 }
 
 export interface InventoryItemResponse extends InventoryItem {
+  _id: string
   itemDefinition: string
   attributes?: {
     attribute: Attribute
@@ -15,9 +19,13 @@ export interface InventoryItemResponse extends InventoryItem {
   }[]
 }
 
-export type CategoryResponse = Category
+export interface CategoryResponse extends Category {
+  _id: string
+}
 
-export type AttributeResponse = Attribute
+export interface AttributeResponse extends Attribute {
+  _id: string
+}
 
 export type ServerResponse =
   | UserResponse
