+++ utils/validators.ts
@@ -1,21 +1,5 @@
-import { ObjectId } from 'mongodb'
+import { Property, validateProperties } from 'utils/validation'
 import deepCopy from './deepCopy'
-
-export interface Property {
-  key: string
-  types: string
-  required: boolean
-}
-
-export interface ValidationResult {
-  success: boolean
-  message: string
-}
-
-interface ValidationErrorList {
-  errorName: string
-  errors: string[]
-}
 
 /**
  * Checks the validity of an Attribute object. Does not validate child objects.
@@ -28,9 +12,9 @@ export function validateAttributeRequest(
 ) {
   let validationModel: Property[]
   if (requestType === 'PUT') {
-    validationModel = attributePostModelProperties
-  } else if (requestType === 'POST') {
     validationModel = attributePutModelProperties
+  } else if (requestType === 'POST') {
+    validationModel = attributePostModelProperties
   } else {
     validationModel = attributeRequestModelProperties
   }
@@ -48,9 +32,6 @@ export function validateCategoryRequest(
   requestType?: 'PUT' | 'POST'
 ) {
   let validationModel: Property[]
-  console.log(categoryRequestModelProperties)
-  console.log(categoryPostModelProperties)
-  console.log(categoryPutModelProperties)
   if (requestType === 'PUT') {
     validationModel = categoryPutModelProperties
   } else if (requestType === 'POST') {
@@ -73,9 +54,9 @@ export function validateItemDefinitionRequest(
 ) {
   let validationModel: Property[]
   if (requestType === 'PUT') {
-    validationModel = itemDefinitionPostModelProperties
-  } else if (requestType === 'POST') {
     validationModel = itemDefinitionPutModelProperties
+  } else if (requestType === 'POST') {
+    validationModel = itemDefinitionPostModelProperties
   } else {
     validationModel = itemDefinitionRequestModelProperties
   }
@@ -94,9 +75,9 @@ export function validateInventoryItemRequest(
 ) {
   let validationModel: Property[]
   if (requestType === 'PUT') {
-    validationModel = inventoryItemPostModelProperties
-  } else if (requestType === 'POST') {
     validationModel = inventoryItemPutModelProperties
+  } else if (requestType === 'POST') {
+    validationModel = inventoryItemPostModelProperties
   } else {
     validationModel = inventoryItemRequestModelProperties
   }
@@ -123,101 +104,6 @@ export function validateUserRequest(
   }
 
   return validateProperties(validationModel, user)
-}
-
-/**
- * Validates a given id to ensure it is a valid ObjectId. Throws an error if invalid.
- * @param id - The ObjectId to validate
- * @returns true if the objectId was valid, false if not
- */
-export function validateObjectId(id: string) {
-  if (!id) return false
-  if (ObjectId.isValid(id)) {
-    if (String(new ObjectId(id)) !== id) return false
-    return true
-  }
-  return false
-}
-
-// TODO improve array validation using 'instanceof Array'
-/**
- * Generic object validator that checks for required fields, extra fields, and type mismatches
- * @param modelProperties An array of Property objects
- * @param obj The object to validate
- * @returns A ValidationResult object
- */
-export function validateProperties(
-  modelProperties: Property[],
-  obj: Record<string, unknown>
-): ValidationResult {
-  const result: ValidationResult = {
-    success: true,
-    message: '',
-  }
-
-  const missingProperties: string[] = []
-  const invalidProperties: string[] = []
-  const typeMismatches: string[] = []
-  const validationErrors: ValidationErrorList[] = [
-    { errorName: 'Missing Required Properties:', errors: missingProperties },
-    { errorName: 'Invalid Properties:', errors: invalidProperties },
-    { errorName: 'Type Mismatches:', errors: typeMismatches },
-  ]
-
-  if (typeof obj !== 'object') {
-    result.message = 'Validation target is not an instance of an object.'
-    result.success = false
-    return result
-  }
-
-  // ensures all required properties are present
-  const isValidObject = true
-  for (const modelProp of modelProperties) {
-    if (
-      isValidObject &&
-      modelProp.required &&
-      !obj.hasOwnProperty(modelProp.key)
-    ) {
-      missingProperties.push(modelProp.key)
-    }
-  }
-
-  // checks validity of all object properties.
-  for (const objKey in obj) {
-    const i = modelProperties.findIndex((prop) => prop.key === objKey)
-
-    // property is not part of server model
-    if (i < 0) {
-      invalidProperties.push(objKey)
-      continue
-    }
-
-    // type mismatch
-    if (!modelProperties[i].types.includes(typeof obj[objKey])) {
-      typeMismatches.push(
-        `'${objKey}': Expected type '${
-          modelProperties[i].types
-        }' but got type '${typeof obj[objKey]}'`
-      )
-    }
-  }
-  for (const errorList of validationErrors) {
-    if (errorList.errors.length) {
-      result.success = false
-      result.message += errorList.errorName
-      let firstError = true
-      for (const error of errorList.errors) {
-        if (!firstError) {
-          result.message += ','
-        }
-        result.message += ` ${error}`
-        firstError = false
-      }
-      result.message += '\n'
-    }
-  }
-  console.log(result)
-  return result
 }
 
 /*
